+++ Localization/enUS.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-11-20 16-07-08 ]] --
+-- [[ Exported at 2022-11-20 23-40-57 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --